pyproject.toml
[project]
name = "tg-content-downloader"
version = "0.1.0"
description = "Telegram bot for downloading content from Youtube/Instagram"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aio-pika>=9.5.8",
    "aiogram>=3.24.0",
    "asyncpg>=0.31.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.46",
]
